// src/data/api/pla-wishlist-data.ts
export interface UserError {
  code: string;
  message: string;
}

export interface DiscoveredProduct {
  sku: string;
  name: string;
  __typename: string;
}

export interface WishlistItemShape {
  id: string;
  product: {
    sku: string;
    name: string;
    __typename: string;
  };
  quantity: number;
  __typename: string;
}

export interface WishlistShape {
  id: string;
  items_count: number;
  items_v2: {
    items: WishlistItemShape[];
  };
  __typename: string;
}

export interface PlaWishlistDataShape {
  nonExistentSku: string;
  nonExistentWishlistItemId: string;
  productSearchTerm: string;
  wishlistItemQuantity: number;
}

export interface PlaWishlistErrorCategories {
  unauthorized: string;
}

export const plaWishlistData: PlaWishlistDataShape = {
  nonExistentSku: 'INVALID-SKU-WISHLIST-TEST-99999',
  nonExistentWishlistItemId: '99999999',
  productSearchTerm: 'shoe',
  wishlistItemQuantity: 1,
};

export const plaWishlistErrorCategories: PlaWishlistErrorCategories = {
  unauthorized: 'graphql-authorization',
};